refactor: extract env vars to a file
