+++ utils/crypto/setEnv.sh
